Rename package to ftw.structlog.
