+++ docs/LICENSE.txt
@@ -1,4 +1,4 @@
-  ftw.jsonlog is copyright 4teamwork AG
+  ftw.structlog is copyright 4teamwork AG
 
   This program is free software; you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
