+++ ftw/structlog/demo/configure.zcml
@@ -3,7 +3,7 @@
     xmlns:browser="http://namespaces.zope.org/browser"
     xmlns:i18n="http://namespaces.zope.org/i18n"
     xmlns:plone="http://namespaces.plone.org/plone"
-    i18n_domain="ftw.jsonlog.demo">
+    i18n_domain="ftw.structlog.demo">
 
     <browser:page
         for="*"
