+++ test-plone-4.3.x.cfg
@@ -2,4 +2,4 @@
 extends =
     https://raw.github.com/4teamwork/ftw-buildouts/master/test-plone-4.3.x.cfg
 
-package-name = ftw.jsonlog
+package-name = ftw.structlog
